core: add support for macos builds

+++ bin/deploy.sh
@@ -8,9 +8,11 @@ log() {
 
 log "Ensuring artifact folders are present..."
 test -e artifacts/linux-x86-64 || exit 2
+test -e artifacts/macos-x86-64 || exit 2
 
 log "Copying artifacts into their respective packages..."
 cp artifacts/linux-x86-64/lib/libsqlite3.so.0.8.6 libsqlite3-x86_64-linux/libsqlite3.so
+cp artifacts/macos-x86-64/lib/libsqlite3.0.dylib libsqlite3-x86_64-linux/libsqlite3.dylib
 
 log "Decrypting deploy key..."
 gpg -q \
@@ -24,7 +26,7 @@ chmod 0600 deploy-key
 trap "rm -f deploy-key" EXIT
 
 log "Building packages..."
-for package in "libsqlite3-x86_64-linux"; do
+for package in "libsqlite3-x86_64-linux" "libsqlite3-x86_64-macosx"; do
     log "Building '$package'..."
     pushd "$package"
